Started refactoring to remove the old single api method
