+++ API/Constants/CacheProfiles.cs
@@ -15,12 +15,12 @@ public static class EasyCacheProfiles
     /// Cache the libraries on the server
     /// </summary>
     public const string Library = "library";
-    /// <summary>
-    /// Metadata filter
-    /// </summary>
-    public const string Filter = "filter";
     public const string KavitaPlusReviews = "kavita+reviews";
     public const string KavitaPlusRecommendations = "kavita+recommendations";
     public const string KavitaPlusRatings = "kavita+ratings";
     public const string KavitaPlusExternalSeries = "kavita+externalSeries";
+    /// <summary>
+    /// Series Detail page for Kavita+ stuff
+    /// </summary>
+    public const string KavitaPlusSeriesDetail = "kavita+seriesDetail";
 }
