+++ API/Controllers/ServerController.cs
@@ -272,6 +272,8 @@ public class ServerController : BaseApiController
         await provider.FlushAsync();
         provider = _cachingProviderFactory.GetCachingProvider(EasyCacheProfiles.KavitaPlusExternalSeries);
         await provider.FlushAsync();
+        provider = _cachingProviderFactory.GetCachingProvider(EasyCacheProfiles.KavitaPlusSeriesDetail);
+        await provider.FlushAsync();
         return Ok();
     }
