+++ API/Extensions/ApplicationServiceExtensions.cs
@@ -85,13 +85,13 @@ public static class ApplicationServiceExtensions
             options.UseInMemory(EasyCacheProfiles.License);
             options.UseInMemory(EasyCacheProfiles.Library);
             options.UseInMemory(EasyCacheProfiles.RevokedJwt);
-            options.UseInMemory(EasyCacheProfiles.Filter);
 
             // KavitaPlus stuff
             options.UseInMemory(EasyCacheProfiles.KavitaPlusReviews);
             options.UseInMemory(EasyCacheProfiles.KavitaPlusRecommendations);
             options.UseInMemory(EasyCacheProfiles.KavitaPlusRatings);
             options.UseInMemory(EasyCacheProfiles.KavitaPlusExternalSeries);
+            options.UseInMemory(EasyCacheProfiles.KavitaPlusSeriesDetail);
         });
 
         services.AddMemoryCache(options =>
